requirements.txt
bcrypt>=4.1.2 
fastapi>=0.109.02
sqlAlchemy>=2.0.27
python-jose[cryptography]>=3.3.0
uvicorn[standard]>=0.27.1
python-dotenv>=1.0.1
python-multipart>=0.0.9
aiofiles>=23.2.1
